fix: 메인 폴더 폴더링 버그 해결
