+++ iOS_Wegg/iOS_Wegg/Models/EmojiModel/EmojiModel.swift
