+++ iOS_Wegg/iOS_Wegg/Models/HotPlaceModel/HotPlaceSectionModel.swift
