+++ iOS_Wegg/iOS_Wegg/Models/NetworkModel/Map/HotPlaceRequest.swift
