+++ iOS_Wegg/iOS_Wegg/Models/NetworkModel/Post/UploadRequest.swift
