+++ iOS_Wegg/iOS_Wegg/Models/NetworkModel/Schedule/ScheduleSearchRequest.swift
