+++ iOS_Wegg/iOS_Wegg/Models/NetworkModel/SignUp/BaseResponse.swift
