+++ iOS_Wegg/iOS_Wegg/Models/NetworkModel/SignUp/SignUpResponse.swift
