+++ iOS_Wegg/iOS_Wegg/Models/NetworkModel/User/UserOccupation.swift
