+++ iOS_Wegg/iOS_Wegg/Models/Protocol/CameraSessionManagerDelegate.swift
