+++ iOS_Wegg/iOS_Wegg/Models/Protocol/MapManagerProtocol.swift
