+++ iOS_Wegg/iOS_Wegg/Models/Todo/TodoRequest.swift
@@ -1,0 +1,13 @@
+//
+//  TodoRequest.swift
+//  iOS_Wegg
+//
+//  Created by KKM on 2/16/25.
+//
+
+import Foundation
+
+struct TodoRequest: Encodable {
+    let status: String
+    let content: String
+}
